backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==1.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0 